requirements.txt
# Core dependencies
numpy>=1.19.0
pandas>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
biopython>=1.78
scikit-learn>=0.24.0
faiss-cpu>=1.7.0
sentence-transformers>=2.0.0
tqdm>=4.50.0
requests>=2.25.0
openai>=1.0.0